[feature/ARCHITECTURE-DESIGN]: Architecture Design Project
* Design library structure, file naming conventions, configurations for health checks and exception handler.
