+++ CSharp.Beginner.Microservice.Restful/Models/UserModel.cs
@@ -1,6 +1,6 @@
 ﻿using System;
-using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
+using System.ComponentModel.DataAnnotations;
 
 namespace CSharp.Beginner.Microservice.Restful.Models;
 
@@ -10,7 +10,7 @@ namespace CSharp.Beginner.Microservice.Restful.Models;
  *  Author   Cristian Hume Henriquez
  *  Version  1.0.0
  *  Created  2022-11-27
- *  Modified 2022-11-27 
+ *  Modified 2022-11-28 
  */
 public class UserModel
 {
@@ -48,6 +48,7 @@ public class UserModel
         get { return _id; }
         set { _id = value; }
     }
+
     [Required]
     [JsonPropertyName("first_name")]
     public String FirstName
@@ -55,6 +56,7 @@ public class UserModel
         get { return _firstName; }
         set { _firstName = value; }
     }
+
     [Required]
     [JsonPropertyName("last_name")]
     public String LastName
@@ -62,6 +64,7 @@ public class UserModel
         get { return _lastName; }
         set { _lastName = value; }
     }
+
     [Required]
     [JsonPropertyName("nationality")]
     public String Nationality
@@ -69,6 +72,7 @@ public class UserModel
         get { return _nationality; }
         set { _nationality = value; }
     }
+
     [Required]
     [JsonPropertyName("occupation")]
     public String Occupation
@@ -76,24 +80,28 @@ public class UserModel
         get { return _occupation; }
         set { _occupation = value; }
     }
+
     [JsonPropertyName("phone")]
     public String Phone
     {
         get { return _phone; }
         set { _phone = value; }
     }
+
     [JsonPropertyName("email")]
     public String Email
     {
         get { return _email; }
         set { _email = value; }
     }
+
     [JsonPropertyName("known_for")]
     public List<String> KnownFor
     {
         get { return _knownFor; }
         set { _knownFor = value; }
     }
+
     [JsonPropertyName("awards")]
     public List<String> Awards
     {
